reduced func complexity to fix lint error
